refactor(UserTestGroups): Split out getTestGroups for easier testability / mocking
