+++ assets/src/components/map.tsx
@@ -44,7 +44,7 @@ import garages, { Garage } from "../data/garages"
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import garageIcon from "../../static/images/icon-bus-garage.svg"
-import inTestGroup, { MAP_BETA_GROUP_NAME } from "../userTestGroups"
+import inTestGroup, { MAP_BETA_GROUP_NAME } from "../userInTestGroup"
 
 export interface Props {
   vehicles: Vehicle[]
