+++ assets/tests/userTestGroups.test.ts
@@ -1,31 +1,19 @@
 import { StructError } from "superstruct"
 import appData from "../src/appData"
-import inTestGroup from "../src/userTestGroups"
+import getTestGroups from "../src/userTestGroups"
 
 jest.mock("appData")
 
-describe("User Test Groups API", () => {
-  test("returns true if user is in test group", () => {
+describe("getTestGroups", () => {
+  test("returns all tests groups when properly formatted", () => {
     const group_1 = "user-test-group-1"
     const group_2 = "user-test-group-2"
     ;(appData as jest.Mock).mockImplementation(() => ({
       userTestGroups: JSON.stringify([group_1, group_2]),
     }))
 
-    expect(inTestGroup(group_1)).toBe(true)
-    expect(inTestGroup(group_2)).toBe(true)
-    expect(inTestGroup(group_1 + group_2)).toBe(false)
+    expect(getTestGroups()).toEqual([group_1, group_2])
   })
-
-  test.each([undefined, {}, { userTestGroup: undefined }])(
-    "returns false if the test group information is not found: case %#",
-    (mock) => {
-      // Missing data key
-      ;(appData as jest.Mock).mockImplementation(() => mock)
-
-      expect(inTestGroup("non-existent-group")).toBe(false)
-    }
-  )
 
   test.each([
     // JSON.parse should throw when given invalid json
@@ -52,9 +40,7 @@ describe("User Test Groups API", () => {
       ;(appData as jest.Mock).mockImplementationOnce(() => ({
         userTestGroups: mockTestGroupData,
       }))
-      expect(() => inTestGroup("non-existent-test-group")).toThrowError(
-        expectedError
-      )
+      expect(() => getTestGroups()).toThrowError(expectedError)
     }
   )
 })
