+++ assets/tests/util/mapMode.test.tsx
@@ -1,17 +1,18 @@
 import React from "react"
-import appData from "../../src/appData"
 import MapPage from "../../src/components/mapPage"
 import SearchPage from "../../src/components/searchPage"
-import { MAP_BETA_GROUP_NAME } from "../../src/userTestGroups"
+import { MAP_BETA_GROUP_NAME } from "../../src/userInTestGroup"
+import getTestGroups from "../../src/userTestGroups"
 import { mapModeForUser } from "../../src/util/mapMode"
 
-jest.mock("appData")
+jest.mock("userTestGroups", () => ({
+  __esModule: true,
+  default: jest.fn(() => []),
+}))
 
 describe("mapModeForUser", () => {
   test("returns new map mode when a part of new map test group", () => {
-    ;(appData as jest.Mock).mockImplementationOnce(() => ({
-      userTestGroups: JSON.stringify([MAP_BETA_GROUP_NAME]),
-    }))
+    ;(getTestGroups as jest.Mock).mockReturnValueOnce([MAP_BETA_GROUP_NAME])
     expect(mapModeForUser()).toEqual({
       path: "/map",
       title: "Map",
@@ -20,9 +21,8 @@ describe("mapModeForUser", () => {
   })
 
   test("returns old search mode when not a part of new map test group", () => {
-    ;(appData as jest.Mock).mockImplementationOnce(() => ({
-      userTestGroups: JSON.stringify([]),
-    }))
+    ;(getTestGroups as jest.Mock).mockReturnValueOnce([])
+
     expect(mapModeForUser()).toEqual({
       path: "/search",
       title: "Search",
